Extract KissCel class to own module
