+++ javascript/kiss.html
@@ -227,6 +227,6 @@
 
     <script src="setdata.js" type="text/javascript"></script>
 
-    <script src="doll.js" type="text/javascript"></script>
+    <script src="doll.js" type="module" type="text/javascript"></script>
   </body>
 </html>
